+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- improvements to the code moving interactive logic of the "in" command into `cmd/common.go`
+- "in clone" is now interactive and will ask the user to confirm the time entry data before
+  creating it.
+- minor grammar bug fixes
+
 ## [v0.2.2] - 2020-03-18
 
 ## Fixed